FEAT. 자유게시판 게시글 상세조회, 좋아요 기능 구현

+++ Devlog/src/main/java/com/devlog/project/board/freeboard/model/dto/Freeboard.java
@@ -11,7 +11,7 @@ import lombok.ToString;
 @Setter
 @ToString
 public class Freeboard {
-	private int boardNo;
+	private Long boardNo;
 	private String boardTitle;
 	private String boardContent;
 	private String bCreateDate; // LocalDateTime?
@@ -27,7 +27,7 @@ public class Freeboard {
 	
 	
 	// 회원 join
-	private int memberNo;
+	private Long memberNo;
 	private String memberNickname; 
 	private String profileImg;
 	private String thumbnail;
